+++ pixoapi/views/categories_view.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from pixoapi.models import Category
+
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields = ['id', 'label']
