+++ pixoapi/views/image_view.py
@@ -1,0 +1,9 @@
+
+from rest_framework import serializers
+from pixoapi.models import Image
+
+
+class ImageSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Image
+        fields = ['id', 'img_url']
